Organize file in packages, data and service. Data contains data classes and service contains interface for endpoint and Retrofit. (TT-101)
