+++ Login-UsernamePassword/src/main/java/hr/foi/techtitans/ttpay/login_UsernamePassword/data/LoginRequest-UserPass.kt
@@ -1,4 +1,4 @@
-package hr.foi.techtitans.ttpay.login_UsernamePassword.data.service
+package hr.foi.techtitans.ttpay.login_UsernamePassword.data
 
 data class LoginRequestData(
     val username: String?,
