+++ Login-UsernamePassword/src/main/java/hr/foi/techtitans/ttpay/login_UsernamePassword/service/LoginService-UserPass.kt
@@ -1,5 +1,7 @@
-package hr.foi.techtitans.ttpay.login_UsernamePassword.data.service
+package hr.foi.techtitans.ttpay.login_UsernamePassword.service
 
+import hr.foi.techtitans.ttpay.login_UsernamePassword.data.LoginRequestData
+import hr.foi.techtitans.ttpay.login_UsernamePassword.data.LoginResponseData
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.POST
